remove bug position.cpp

+++ projet_ballon/software/lib/Position/Position.h
@@ -16,8 +16,8 @@
 class Position {
 public:
     
-    Position(const double _latitude, 
-             const double _longitude, 
+    Position(const float _latitude, 
+             const float _longitude, 
              String _comment = "", 
              const char _symboleTable = '/',
              const char _symbole= '>'
@@ -27,10 +27,10 @@ public:
     virtual ~Position();
     
     char* getPduAprs(bool compressed = false);
-    void setLatitude(const double _latitude);
-    void setLongitude(const double _longitude);
+    void setLatitude(const float _latitude);
+    void setLongitude(const float _longitude);
     void setComment(String _comment);
-    void setAltitude(const double _alt);
+    void setAltitude(const float _alt);
     
 protected:
     void latitude_to_str();
@@ -38,8 +38,8 @@ protected:
     void convBase91(int x, char* base91);
     char    slat[9];
     char    slong[10];
-    double  latitude;   
-    double  longitude;
+    float   latitude;   
+    float   longitude;
     char    symboleTable;
     char    symbole;
     char    pdu[100];
